# internal/admin/openapi.json
{
  "components": {
    "schemas": {
      "Audit": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "time": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Fresh TOTP or unused recovery code. Development mode permits an empty value.",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "code"
        ],
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "development": {
            "type": "boolean"
          },
          "expires": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user",
          "csrf_token",
          "development"
        ],
        "type": "object"
      },
...
      "Success": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "warning": {
            "type": "string"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "created": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "operator",
              "viewer"
            ],
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "csrf": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "session": {
        "in": "cookie",
        "name": "__Host-waf_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "All paths are relative to /api/v1 on the dedicated administration domain. Mutations require the session cookie, X-CSRF-Token and exact Origin. Body limit: 2 MiB. Sites and global security policies are edited together through the versioned configuration draft. Development mode uses the waf_session_dev cookie on loopback HTTP.",
    "title": "WAF Administration API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/audit": {
      "get": {
        "operationId": "get_audit",
        "parameters": [
          {
            "description": "Exclusive event/audit ID cursor; omit for newest records.",
            "in": "query",
            "name": "before",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Audit"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Read administration audit trail.",
        "tags": [
          "audit"
        ],
        "x-required-role": "viewer"
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "post_auth_login",
        "parameters": [
          {
            "description": "Must match the exact console origin.",
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          }
        },
        "security": [],
        "summary": "Authenticate using password and TOTP; sets an HttpOnly session cookie.",
        "tags": [
          "auth"
        ],
        "x-required-role": "anonymous"
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "post_auth_logout",
        "parameters": [
          {
            "description": "Must match the exact console origin.",
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request rejected or operation failed; see error."
          }
        },
        "security": [
          {
            "csrf": [],
            "session": []
          }
        ],
        "summary": "Revoke current session.",
        "tags": [
          "auth"
        ],
        "x-required-role": "viewer"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}